ENCD-5367 clean up body map styles
