+++ src/encoded/static/scss/style.scss
@@ -239,6 +239,7 @@ $link-hover-color:                          darken($link-color, 15%);
         "encoded/modules/search",
         "encoded/modules/summary",
         "encoded/modules/body_map",
+        "encoded/modules/body_map_facet",
         "encoded/modules/genome_browser",
         "encoded/modules/characterizations",
         "encoded/modules/audits",
